clickshop-demo/requirements.txt
# Strands Framework
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# AWS Services
boto3>=1.34.0
botocore>=1.34.0

# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6  # For file uploads

# Vector operations (for pgvector result parsing)
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# For pretty output
rich>=13.0.0
tabulate>=0.9.0

# Testing
pytest>=7.4.0
hypothesis>=6.100.0